consumer/requirements.txt
pika==1.2.0
Pillow==9.0.1